Adds --map-hash option for Hash-like objects
jsontool works very well with arrays and record-like objects with fixed keys.
However, in many cases, the object is being used as a Hash and the keys will not
be known ahead of time. Moreover, we might want to extract and print the key
value in the final lookup results.

This option tells the tool to remap the object as an array of objects with keys
"key" and "value", where "key" is the old key and "value" is the JSON value that
it pointed to.